Safari-609.1.11.1/LayoutTests/js/dom/domjit-accessor-effect-should-overlap-with-call-expected.txt

Test function calls should overlap with effects of DOMJIT accessors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test() is true
PASS successfullyParsed is true

TEST COMPLETE
